HW 11, t3, p.1

+++ src/lesson11/fruits/Fruit.java
@@ -1,9 +1,9 @@
 package lesson11.fruits;
 
 public class Fruit {
-    protected int weight;
+    protected double weight;
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
